requirements.txt
Flask==2.2.5
python-dotenv
pydantic==1.9.1
PyJWT==2.4.0
Werkzeug==2.2.2
Flask-Cors==5.0.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
pydantic[email]==1.9.1
pytest==7.2.2
Flask-Testing==0.8.1
gunicorn==20.1.0
psycopg2-binary
resend
celery
redis==4.5.1
pandas==1.5.3
flasgger==0.9.7b2
flask-cors
requests
google-auth
google-auth-oauthlib